Move CollapsibleBox and DetachedPanelWindow to graphical/qt_widgets.py
qt_gui.py split, second cut. Both classes move verbatim; MainWindow is now
referenced only under TYPE_CHECKING so the runtime import direction stays
one-way (qt_gui -> qt_widgets). qt_gui re-exports both names for backward
compatibility.
